# Add thirdparty library etcpak for faster imports.