… COPYRIGHT.txt
… modules/etc/SCsub
… modules/etc/image_compress_etc.cpp
… modules/etc/image_compress_etc.h
… modules/etc/register_types.cpp
… modules/etc/register_types.h
… modules/etc/texture_loader_pkm.cpp
… modules/etc/texture_loader_pkm.h
… modules/etcpak/config.py
… modules/squish/image_compress_squish.cpp
… thirdparty/README.md
… thirdparty/etc2comp/Etc.cpp
… thirdparty/etc2comp/Etc.h
… thirdparty/etc2comp/EtcBlock4x4.cpp
… thirdparty/etc2comp/EtcBlock4x4.h
… thirdparty/etc2comp/EtcBlock4x4Encoding.cpp
… thirdparty/etc2comp/EtcBlock4x4Encoding.h
… thirdparty/etc2comp/EtcBlock4x4EncodingBits.h
… thirdparty/etc2comp/EtcBlock4x4Encoding_ETC1.cpp
… thirdparty/etc2comp/EtcBlock4x4Encoding_ETC1.h
… thirdparty/etc2comp/EtcBlock4x4Encoding_R11.cpp
… thirdparty/etc2comp/EtcBlock4x4Encoding_R11.h
… thirdparty/etc2comp/EtcBlock4x4Encoding_RG11.cpp
… thirdparty/etc2comp/EtcBlock4x4Encoding_RG11.h
… thirdparty/etc2comp/EtcBlock4x4Encoding_RGB8.cpp
… thirdparty/etc2comp/EtcBlock4x4Encoding_RGB8.h
… thirdparty/etc2comp/EtcBlock4x4Encoding_RGB8A1.cpp
… thirdparty/etc2comp/EtcBlock4x4Encoding_RGB8A1.h
… thirdparty/etc2comp/EtcBlock4x4Encoding_RGBA8.cpp
… thirdparty/etc2comp/EtcBlock4x4Encoding_RGBA8.h