**Add `etcpak` library for faster ETC/ETC2/S3TC imports.**

- `etc` module was renamed to `etcpak` and modified to use the new library.
- PKM importer is removed in the process. Could be re-added if still useful?
- Old library `etc2comp` is removed.
- S3TC compression no longer done via `squish` (but decompression still is).
- Slight modifications to etcpak sources for MinGW compatibility,
  to fix LLVM `-Wc++11-narrowing` errors, and to allow using vendored or
  system libpng.

----

Squish and etc modules have been modified.

Modifications for mingw.

Some times from TPS sample. 

ook ook ook :banana:

Please help measure if this is faster or not.

```
ETCPAK encode took 74 ms
Encoding format: ETC2_RA_AS_RG
ETCPAK encode took 60 ms
Encoding format: DXT5 RGBA8
ETCPAK encode took 74 ms
Encoding format: ETC2_RGB8
ETCPAK encode took 130 ms
Encoding format: DXT1 RGB8
ETCPAK encode took 7 ms
Encoding format: ETC
ETCPAK encode took 30 ms
Encoding format: FORMAT_DXT5_RA_AS_RG
ETCPAK encode took 9 ms
Encoding format: ETC2_RA_AS_RG
ETCPAK encode took 135 ms
Encoding format: DXT5 RGBA8
ETCPAK encode took 8 ms
Encoding format: ETC2_RGB8
```