@fire 2021-03-26
Test using.

```c++
// https://github.com/godotengine/godot/blob/master/editor/editor_file_system.cpp#L1974
for (int i = 0; i < files.size(); i++) {	
  uint64_t t = OS::get_singleton()->get_ticks_msec();
  pr.step(files[i].path.get_file(), i);
  _reimport_file(files[i].path);
  uint16_t final = OS::get_singleton()->get_ticks_msec() - t;
  print_line(vformat("%s import took %f.001 seconds", files[i].path, rtos(final / 1000.0)));
}
```