@fire 2021-03-26
From a 8k texture load test, 37 seconds is get_pixel and 13 seconds is texture compression (this pr),

https://github.com/fire/texture-import-godot-project/tree/main/peacock