@fire 2021-03-27
Added some flags:

```
env.Prepend(CCFLAGS=["-msse4.1"])
env.Prepend(CCFLAGS=["-mavx"])
env.Prepend(CCFLAGS=["-mavx2"])
env.Prepend(CCFLAGS=["-march=skylake"])
```