@fire 2021-03-30
I don't have the full numbers because there's another pr for better import time logging but here they are.

ETCPAK encode took 920 ms (dxt5)

ETCPAK encode took 3130 ms (etc2)