@akien-mga 2021-04-06
> etc2 import with quicketc2 modified optimization which need to be up streamed.

Does this PR include the QuickETC2 patch? Because that's not documented and it says it's on https://github.com/wolfpld/etcpak/commits/403d38b3f1cb347c196d845d0a05e44a00d17169 which is upstream `master` without QuickETC2.