@fire 2021-04-07
I added the copyright from jtbr's bswap and bswap64.