@fire 2021-04-07
I'll try to get this error fixed upstream.

```
type 'int' to 'uint32_t' type narrowing.
```

https://github.com/wolfpld/etcpak/issues/11