@hpvb 2021-04-12
I was able to build etcpak on mingw with only the following diff:
```
diff --git a/ProcessRGB.cpp b/ProcessRGB.cpp
index 220d5c5..b9e90a8 100644
--- a/ProcessRGB.cpp
+++ b/ProcessRGB.cpp
@@ -1,5 +1,6 @@
 #include <array>
 #include <string.h>
+#include <limits>
 
 #ifdef __ARM_NEON
 #  include <arm_neon.h>
diff --git a/System.cpp b/System.cpp
index 1383d0e..a09b289 100644
--- a/System.cpp
+++ b/System.cpp
@@ -35,7 +35,7 @@ unsigned int System::CPUCores()
 
 void System::SetThreadName( std::thread& thread, const char* name )
 {
-#ifdef _WIN32
+#ifdef _MSC_VER
     const DWORD MS_VC_EXCEPTION=0x406D1388;
 
 #  pragma pack( push, 8 )
```
This was a cross-build so I will try on native windows as well, but I don't think we need quite as many chances to etcpak to make it work for us.