@akien-mga 2021-04-12
@hpvb With your changes (and without https://github.com/godotengine/godot/pull/47370/commits/853b03548de68d66749809df355e56817b74d1d3) I still get an error trying to include `byteswap.h` here:
https://github.com/wolfpld/etcpak/blob/403d38b3f1cb347c196d845d0a05e44a00d17169/ProcessRGB.cpp#L24-L30

Removing this `#else` then works fine as what we want for MinGW/GCC is to use the `__builtin_bswap32` and `__builtin_bswap64` which are aliases right below. Might break other platforms, not so sure what's the exact intent of this code since upstream seems to use `_WIN32` and `_MSC_VER` interchangeably.