@akien-mga 2021-04-12
I made a branch based on this PR, with the second commit removed, and replaced by the above patch (minus the addition of `<limits>` which didn't seem needed in the end?) + some other packaging cleanups. WDYT?