@akien-mga 2021-04-12
I made a version with both commits squashed together and the commit message edited to add some more details (please confirm that I understood and described the changes correctly):

https://github.com/akien-mga/godot/tree/etcpak-cleaned-up

I'll try to do some tests building on Windows with MSVC and MinGW to confirm that it works OK.