@fire 2021-04-12
The commit message is correct.

Minor comments:

1. lz4/lz4.c What's the difference between lz4 and minil4?
2. I also removed an etc framing format -- pkm. I've not heard of people using it.