@akien-mga 2021-04-12
Added `<limits>` back. Updated my branch with further fixes to the `SCsub`: https://github.com/akien-mga/godot/tree/etcpak-cleaned-up

I clarified the dependency on libpng, and remove the two defines (`NO_GZIP` and `PNG_INTEL_SSE`) which would likely conflict with our zlib and libpng build depending on what the build order is (redefining defines which impact a header which has possibly already been parsed likely doesn't work reliably). We can re-evaluate them if they're relevant to enable for the whole engine, but not only for etcpak's sake.

Now I see the narrowing errors on the CI which are specific to LLVM, not to mingw:
https://github.com/akien-mga/godot/runs/2325019211?check_suite_focus=true

So I added back @fire's changes with a dedicated patch (separate from the mingw patch). I don't know if they're the proper way to fix the issue, it's just what LLVM recommends (so I assume it's just making explicit what happens otherwise - doesn't mean that it's not an actual bug that should be fixed upstream).